Add warning about using Vercel
Putting a more prominent warning against using Vercel. Vercel has continued to ignore my requests to not be featured on their webpage (multiple emails to the templates team).

They are copying my work without following the MIT license which requires: "... subject to the following conditions: The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.". See the LICENSE file.

They have refused a copyright takedown request by pointing to the MIT License. However, they not following the singular condition of the MIT license so I don't believe their usage is covered.